Hard Difficulty Fixes
